add tab in branch page
